requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.11
Flask==1.1.2
GitHub-Flask==3.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyGithub==1.54.1
PyJWT==1.7.1
requests==2.25.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1